[layout] Move DstLayout from root, update comments
- Add new `layout` module to hold `DstLayout`
- Remove instance invariants from `DstLayout` and move them to
  `KnownLayout::LAYOUT`
